+++ src/steps/resource-manager/databases/postgresql/index.ts
@@ -1,20 +1,31 @@
 import {
   createDirectRelationship,
+  getRawData,
+  IntegrationStepExecutionContext,
   RelationshipClass,
+  Step,
 } from '@jupiterone/integration-sdk-core';
 
 import { createAzureWebLinker } from '../../../../azure';
-import { IntegrationStepContext } from '../../../../types';
-import { getAccountEntity } from '../../../active-directory';
+import { IntegrationConfig, IntegrationStepContext } from '../../../../types';
+import { getAccountEntity, STEP_AD_ACCOUNT } from '../../../active-directory';
 import { createDatabaseEntity, createDbServerEntity } from '../converters';
 import { PostgreSQLClient } from './client';
-import { PostgreSQLEntities } from './constants';
+import {
+  PostgreSQLEntities,
+  PostgreSQLRelationships,
+  steps,
+} from './constants';
 import createResourceGroupResourceRelationship from '../../utils/createResourceGroupResourceRelationship';
-import { createDiagnosticSettingsEntitiesAndRelationshipsForResource } from '../../utils/createDiagnosticSettingsEntitiesAndRelationshipsForResource';
+import {
+  createDiagnosticSettingsEntitiesAndRelationshipsForResource,
+  diagnosticSettingsEntitiesForResource,
+  diagnosticSettingsRelationshipsForResource,
+} from '../../utils/createDiagnosticSettingsEntitiesAndRelationshipsForResource';
+import { STEP_RM_RESOURCES_RESOURCE_GROUPS } from '../../resources/constants';
+import { Server } from '@azure/arm-postgresql/esm/models';
 
-export * from './constants';
-
-export async function fetchPostgreSQLDatabases(
+export async function fetchPostgreSQLServers(
   executionContext: IntegrationStepContext,
 ): Promise<void> {
   const { instance, logger, jobState } = executionContext;
@@ -23,10 +34,15 @@ export async function fetchPostgreSQLDatabases(
   const client = new PostgreSQLClient(instance.config, logger);
 
   await client.iterateServers(async (server) => {
+    const serverConfigurations = await client.getServerConfigurations({
+      name: server.name!,
+      id: server.id!,
+    });
     const serverEntity = createDbServerEntity(
       webLinker,
       server,
       PostgreSQLEntities.SERVER._type,
+      serverConfigurations,
     );
     await jobState.addEntity(serverEntity);
     await createResourceGroupResourceRelationship(
@@ -38,33 +54,69 @@ export async function fetchPostgreSQLDatabases(
       executionContext,
       serverEntity,
     );
-
-    try {
-      await client.iterateDatabases(server, async (database) => {
-        const databaseEntity = createDatabaseEntity(
-          webLinker,
-          database,
-          PostgreSQLEntities.DATABASE._type,
-        );
-
-        await jobState.addEntity(databaseEntity);
-        await jobState.addRelationship(
-          createDirectRelationship({
-            _class: RelationshipClass.HAS,
-            from: serverEntity,
-            to: databaseEntity,
-          }),
-        );
-      });
-    } catch (err) {
-      logger.warn(
-        { err, server: { id: server.id, type: server.type } },
-        'Failure fetching databases for server',
-      );
-      // In the case this is a transient failure, ideally we would avoid
-      // deleting previously ingested databases for this server. That would
-      // require that we process each server independently, and have a way
-      // to communicate to the synchronizer that only a subset is partial.
-    }
   });
 }
+
+export async function fetchPostgreSQLDatabases(
+  executionContext: IntegrationStepContext,
+): Promise<void> {
+  const { instance, logger, jobState } = executionContext;
+  const accountEntity = await getAccountEntity(jobState);
+  const webLinker = createAzureWebLinker(accountEntity.defaultDomain as string);
+  const client = new PostgreSQLClient(instance.config, logger);
+
+  await jobState.iterateEntities(
+    { _type: PostgreSQLEntities.SERVER._type },
+    async (serverEntity) => {
+      const server = getRawData<Server>(serverEntity);
+
+      if (server) {
+        await client.iterateDatabases(server, async (database) => {
+          const databaseEntity = createDatabaseEntity(
+            webLinker,
+            database,
+            PostgreSQLEntities.DATABASE._type,
+          );
+
+          await jobState.addEntity(databaseEntity);
+          await jobState.addRelationship(
+            createDirectRelationship({
+              _class: RelationshipClass.HAS,
+              from: serverEntity,
+              to: databaseEntity,
+            }),
+          );
+        });
+      }
+    },
+  );
+}
+
+export const postgreSqlSteps: Step<
+  IntegrationStepExecutionContext<IntegrationConfig>
+>[] = [
+  {
+    id: steps.SERVERS,
+    name: 'PostgreSQL Servers',
+    entities: [
+      PostgreSQLEntities.SERVER,
+      ...diagnosticSettingsEntitiesForResource,
+    ],
+    relationships: [
+      PostgreSQLRelationships.RESOURCE_GROUP_HAS_POSTGRESQL_SERVER,
+      ...diagnosticSettingsRelationshipsForResource,
+    ],
+    dependsOn: [STEP_AD_ACCOUNT, STEP_RM_RESOURCES_RESOURCE_GROUPS],
+    executionHandler: fetchPostgreSQLServers,
+  },
+  {
+    id: steps.DATABASES,
+    name: 'PostgreSQL Databases',
+    entities: [PostgreSQLEntities.DATABASE],
+    relationships: [
+      PostgreSQLRelationships.POSTGRESQL_SERVER_HAS_POSTGRESQL_DATABASE,
+    ],
+    dependsOn: [STEP_AD_ACCOUNT, steps.SERVERS],
+    executionHandler: fetchPostgreSQLDatabases,
+  },
+];